Add files via upload
The two orange wires connect the cathode (long legged/left side) of the LED pins to the Raspberry Pi. The two resistors connect the anode (short legged/right side) of the LED pins to the "-" or blue rail of the breadboard, which is also being connect to the "ground" pin on the raspberry pi by the gray wire.