Add files via upload
The orange wires connect the cathode (long legged) side of the LED pins directly to the raspberry pi. The gray wire connects the "-" or blue rail of the breadboard to the ground pin on the raspberry pi. The resistors connect the anode (short legged) side of the LED pins to the blue rail.